+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0]
+
+### Changed
+
+- when in dev mode (`"jsx": "react-jsxdev"`) add a `jsx-src` attribute to every
+  element providing the file url with line and column numbers of the element
+  with the source. The `JSX_SRC_ATTR` env var can be set to override the
+  attribute name or suppress it if set to empty.
+- update dependencies
+
 ## [0.3.0]
 
 ### Changed